requirements.txt
click
boto3
pyyaml